[chore] 전체 엔티티 entity 폴더로 이동  #20
