+++ src/main/java/refooding/api/domain/recipe/entity/Manual.java
@@ -1,4 +1,4 @@
-package refooding.api.domain.recipe;
+package refooding.api.domain.recipe.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -6,21 +6,19 @@ import refooding.api.common.domain.BaseTimeEntity;
 
 @Entity
 @Getter
-public class RecipeIngredient extends BaseTimeEntity {
+public class Manual extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ingredient_id")
-    private Ingredient ingredient;
-
-    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "recipe_id")
     private Recipe recipe;
 
-    private String quantity;
+    private int seq;
 
-    private boolean isMainIngredient;
+    private String content;
+
+    private String imageSrc;
 }
