+++ src/main/java/refooding/api/domain/recipe/entity/RecipeIngredient.java
@@ -1,4 +1,4 @@
-package refooding.api.domain.recipe;
+package refooding.api.domain.recipe.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -6,19 +6,21 @@ import refooding.api.common.domain.BaseTimeEntity;
 
 @Entity
 @Getter
-public class Manual extends BaseTimeEntity {
+public class RecipeIngredient extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "ingredient_id")
+    private Ingredient ingredient;
+
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "recipe_id")
     private Recipe recipe;
 
-    private int seq;
+    private String quantity;
 
-    private String content;
-
-    private String imageSrc;
+    private boolean isMainIngredient;
 }
